removing eds store from ipld-getter
