+++ share/getters/ipld.go
@@ -2,10 +2,10 @@ package getters
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"sync"
+	"sync/atomic"
 
-	"github.com/filecoin-project/dagstore"
 	"github.com/ipfs/go-blockservice"
 	"github.com/ipfs/go-cid"
 	"golang.org/x/sync/errgroup"
@@ -26,23 +26,13 @@ var _ share.Getter = (*IPLDGetter)(nil)
 type IPLDGetter struct {
 	rtrv  *eds.Retriever
 	bServ blockservice.BlockService
-
-	// store is an optional eds store that can be used to cache retrieved shares.
-	store *eds.Store
 }
 
 // NewIPLDGetter creates a new share.Getter that retrieves shares from the IPLD network.
 func NewIPLDGetter(bServ blockservice.BlockService) *IPLDGetter {
-	return NewIPLDGetterWithStore(bServ, nil)
-}
-
-// NewIPLDGetterWithStore creates a new IPLDGetter with an EDS store for storing EDSes from
-// GetShares.
-func NewIPLDGetterWithStore(bServ blockservice.BlockService, store *eds.Store) *IPLDGetter {
 	return &IPLDGetter{
 		rtrv:  eds.NewRetriever(bServ),
 		bServ: bServ,
-		store: store,
 	}
 }
 
@@ -62,14 +52,6 @@ func (ig *IPLDGetter) GetShares(ctx context.Context, root *share.Root) ([][]shar
 	eds, err := ig.rtrv.Retrieve(ctx, root)
 	if err != nil {
 		return nil, fmt.Errorf("getter/ipld: failed to retrieve eds: %w", err)
-	}
-
-	// store the EDS if a store was provided
-	if ig.store != nil {
-		err = ig.store.Put(ctx, root.Hash(), eds)
-		if err != nil && !errors.Is(err, dagstore.ErrShardExists) {
-			return nil, fmt.Errorf("getter/ipld: failed to cache retrieved shares: %w", err)
-		}
 	}
 
 	origWidth := int(eds.Width() / 2)
@@ -131,4 +113,38 @@ func (ig *IPLDGetter) GetSharesByNamespace(
 	}
 
 	return shares, nil
+}
+
+// ses is a struct that can optionally be passed by context to the share.Getter methods using
+// WithSession to indicate that a blockservice session should be created.
+type ses struct {
+	sync.Mutex
+	atomic.Pointer[blockservice.Session]
+}
+
+// WithSession stores an empty ses in the context, indicating that a blockservice session should be
+// created.
+func WithSession(ctx context.Context) context.Context {
+	return context.WithValue(ctx, ses{}, &ses{})
+}
+
+func getGetter(ctx context.Context, service blockservice.BlockService) blockservice.BlockGetter {
+	s, ok := ctx.Value(ses{}).(*ses)
+	if !ok {
+		return service
+	}
+
+	val := s.Load()
+	if val != nil {
+		return val
+	}
+
+	s.Lock()
+	defer s.Unlock()
+	val = s.Load()
+	if val == nil {
+		val = blockservice.NewSession(ctx, service)
+		s.Store(val)
+	}
+	return val
 }
